alsa 告警消除

+++ src/13alsa/13-4mixer/mixer.c
@@ -4,7 +4,7 @@
 int main(void)
 {
 	int unmute, chn;
-	int al, ar, idx = 0;
+	long al, ar, idx = 0;
 	snd_mixer_t *mixer;
 	snd_mixer_elem_t * elem;
 	snd_mixer_open(&mixer, 0);
@@ -18,8 +18,8 @@ int main(void)
 		snd_mixer_selem_get_playback_volume(elem, SND_MIXER_SCHN_FRONT_LEFT, &al);
 		snd_mixer_selem_get_playback_volume(elem, SND_MIXER_SCHN_FRONT_RIGHT, &ar);
 		/*平均音量 */
-		printf("al[%d] volume is %d\n", ++idx, al);
-		printf("ar[%d] volume is %d\n", idx, ar);
+		printf("al[%ld] volume is %ld\n", ++idx, al);
+		printf("ar[%ld] volume is %ld\n", idx, ar);
 		/*设定左右声道的音量 */
 		snd_mixer_selem_set_playback_volume(elem, SND_MIXER_SCHN_FRONT_LEFT, 49);
 		snd_mixer_selem_set_playback_volume(elem, SND_MIXER_SCHN_FRONT_RIGHT, 49);
